requirements.txt
Django==3.1
Pillow==7.2.0
psycopg2-binary==2.8.5
celery==4.4.7

djangorestframework==3.12.1
markdown==3.1.1
django-filter==2.4.0
pyyaml>=5.1
uritemplate==3.0.1

gunicorn==19.9.0

redis==3.5.3

dj-database-url==0.5.0
python-json-logger==0.1.11

django-cors-headers==3.5.0

PyJWT==1.7.1
djangorestframework_simplejwt==4.4.0
